+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Panoramas are now saved as a single file instead of multiple.
 * There are now more toasts for certain errors.
 * You can now enable a button to capture an in-game screenshot from the game menu.
+* Panorama super-sampling is now configurable (Defaults to 4x).
 
 ### Changes
 * **Moved the default unified screenshot directory to match other apps.**
@@ -33,3 +34,6 @@ All translation changes have only made it to the en_us language file at this tim
 - Added `text.snapper.empty.panorama`, to display specific information for empty panorama galleries.
 - Added `button.snapper.helper.screenshot` for a button in the Game Menu which lets users take screenshots without using the keyboard.
 - Added `subtitles.snapper.ui.shutter` for the sound that plays when Capture Screenshot is pressed from the Game Menu.
+- Added `config.snapper.panoramaSuperSampling` for the super sampling slider in the config menu.
+- Added `config.snapper.panoramaSuperSampling.tooltip` for the tooltip of the super sampling slider in the config menu.
+- Added `config.snapper.panoramaSuperSampling.value` for the value of the super sampling slider in the config menu.